streamlit-nutricional/requirements.txt
streamlit 
openai 
guardrails-ai
codecov
python-dotenv
Pillow

